// src/models/IoK8sApiFlowcontrolV1LimitResponse/types.ts
import { IoK8sApiFlowcontrolV1QueuingConfigurationModelOptions as IoK8sApiFlowcontrolV1QueuingConfiguration } from '../IoK8sApiFlowcontrolV1QueuingConfiguration';

export interface IoK8sApiFlowcontrolV1LimitResponseModelOptions {
  queuing?: IoK8sApiFlowcontrolV1QueuingConfiguration;
  type: string;
}